kurerni ham boshlangich tugatdim

+++ src/main/java/uz/nonvoyxona/app/repository/DeliveryRepo.java
@@ -3,5 +3,5 @@ package uz.nonvoyxona.app.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import uz.nonvoyxona.app.model.Delivery;
 
-public interface DeliveryRepo extends JpaRepository<Delivery, Long> {
+public interface DeliveryRepo extends JpaRepository<Delivery, Integer> {
 }
